fix: split search serializers into one file per class for Zeitwerk
Zeitwerk requires each file to define exactly one constant matching its
filename. The combined search_serializer.rb (ExactSearch, FuzzySearch,
NullSearch in one file) violated this and caused eager_load! to raise
Zeitwerk::NameError in CI.
