+++ app/serializers/api/v3/exact_search_serializer.rb
@@ -1,0 +1,12 @@
+module Api
+  module V3
+    class ExactSearchSerializer
+      def self.call(presenter)
+        {
+          type: presenter.type,
+          entry: presenter.entry,
+        }
+      end
+    end
+  end
+end
